Reuse alert component for strip prompt
