+++ src/prompts/alert/alert.scss
@@ -11,7 +11,8 @@
     padding: 20px;
 
     display: flex;
-    flex-direction: column;
+    flex-direction: row;
+    flex-wrap: wrap;
 
     font-family: $defaultFont;
 
@@ -44,23 +45,30 @@
     }
 }
 
+.kumulos-alert-icon {
+    order: 2;
+    width: 70px;
+    height: 70px;
+    align-self: flex-start;
+}
+
+.kumulos-alert-content {
+    display: flex;
+    order: 1;
+    flex-direction: column;
+    flex: 1;
+    margin-right: 10px;
+}
+
 .kumulos-alert-header {
     display: flex;
-    font-size: 24px;
+    font-size: $defaultHeaderFontSize;
+    margin-bottom: 20px;
 
     h1 {
         font-size: inherit;
-        flex: 1;
+        font-weight: 400;
         margin: 0;
-    }
-
-    .kumulos-alert-icon {
-        width: 70px;
-        height: 70px;
-
-        img {
-            width: 100%;
-        }
     }
 }
 
@@ -74,4 +82,8 @@
     display: flex;
     justify-content: flex-end;
     margin-top: 20px;
+    order: 3;
+    width: 100%;
+    height: 30px;
+    align-self: flex-end;
 }
