+++ src/prompts/alert/strip.scss
@@ -1,37 +1,30 @@
 .kumulos-strip-container {
-    background: #fff;
-
-    position: fixed;
-
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    max-height: 100%;
-    padding: 10px;
-
     display: flex;
     flex-direction: row;
+
+    position: fixed;
+    left: 0;
+
+    width: 100%;
+    height: 90px;
+
+    padding: 10px;
+
+    border-color: #fff;
+    background: #fff;
 
     font-family: $defaultFont;
 
     animation: kumulos-anim-fade-in 0.2s ease-in;
     backface-visibility: hidden;
-}
 
-@media only screen and (min-device-width: 480px) {
-    .kumulos-strip-container {
-        height: 90px;
-        border-color: #fff;
+    &.kumulos-prompt-position-top {
+        top: 0;
+    }
 
-        &.kumulos-prompt-position-top {
-            top: 0;
-        }
-
-        &.kumulos-prompt-position-bottom {
-            bottom: 0;
-            top: unset;
-        }
+    &.kumulos-prompt-position-bottom {
+        bottom: 0;
+        top: unset;
     }
 }
 
@@ -39,13 +32,6 @@
     width: 70px;
     height: 70px;
     margin-right: 20px;
-
-    img {
-        max-width: 70px;
-        max-height: 70px;
-        width: auto;
-        height: auto;
-    }
 }
 
 .kumulos-strip-content {
