+++ src/prompts/prompts.scss
@@ -7,4 +7,4 @@
 
 @import './bell/bell.scss';
 @import './alert/alert.scss';
-@import './strip/strip.scss';
+@import './alert/strip.scss';
