+++ types/src/prompts/alert/index.d.ts
@@ -1,7 +1,7 @@
 import { Component, h } from 'preact';
 import { PromptUiProps } from '../ui';
-import { AlertPromptConfig, PlatformConfig } from '../../core';
-export declare class Alert extends Component<PromptUiProps<AlertPromptConfig>, never> {
+import { AlertPromptConfig, PlatformConfig, StripPromptConfig } from '../../core';
+export declare class Alert extends Component<PromptUiProps<AlertPromptConfig | StripPromptConfig>, never> {
     onRequestNativePrompt: () => void;
     onRequestCancel: () => void;
     renderAlert: (platformConfig?: PlatformConfig | undefined) => h.JSX.Element;
